finished rhh master merge

+++ include/field_player_avatar.h
@@ -64,6 +64,6 @@ bool32 IsPlayerSpinExitActive(void);
 void SetPlayerInvisibility(bool8 invisible);
 u8 player_get_pos_including_state_based_drift(s16 *x, s16 *y);
 void StartFishing(u8 rod);
-u8 GetPlayerAvatarGraphicsByGender(u8); //custom PSF function for showing the correct avatar on naming screen based on player's body choice
+u16 GetPlayerAvatarGraphicsByGender(u8); //custom PSF function for showing the correct avatar on naming screen based on player's body choice
 
 #endif // GUARD_FIELD_PLAYER_AVATAR_H
